pyproject.toml
# Build system configuration
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Project metadata
[project]
name = "datapizza-ai"
version = "0.0.9"
description = "Build reliable Gen AI solutions without overhead"
readme = "README.md"
license = {text = "MIT"}

requires-python = ">=3.10.0,<4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "datapizza-ai-core>=0.0.1",
    "datapizza-ai-clients-openai>=0.0.1",
    "datapizza-ai-clients-openai-like>=0.0.1",
    "datapizza-ai-embedders-openai>=0.0.1",
    "datapizza-ai-vectorstores-qdrant>=0.0.1",
    # HTTP client for streaming SSE and MCP connections
    "aiohttp",
    # environment variable loader used in main/orchestrator
    "python-dotenv",
]

# Development dependencies
[dependency-groups]
dev = [
    "deptry>=0.23.0",
    "pytest",
    "ruff>=0.11.5",
]

# UV workspace configuration (removed)
# The original pyproject contained a full datapizza-ai workspace
# declaration.  That made sense for the upstream library repo but
# our hackapizza project is just an application that depends on the
# published packages, it does *not* contain any of the workspace
# members.  When `uv sync` ran it attempted to treat
# ``datapizza-ai-core`` as a local package and failed because that
# directory doesn't exist here.
#
# We can keep the build metadata for the application itself and
# simply remove the workspace blocks.  If you ever do want to work
# on the datapizza libraries themselves, restore the original
# sections or run `uv init` from the appropriate repo.

# (sections deleted)


# Hatch build configuration
[tool.hatch.build.targets.sdist]
include = ["datapizza"]
exclude = ["**/BUILD"]

[tool.hatch.build.targets.wheel]
include = ["datapizza"]
exclude = ["**/BUILD"]

# Ruff configuration
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    # "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
    "C4",  # flake8-comprehensions
]
